feat: add register flow to register page
